chore: move experiment hooks to rleated feature directories
